Creating web service

+++ src/service/doctor_service.py
@@ -27,7 +27,10 @@ def get_one_by_id(doctor_id):
     :param doctor_id:
     :return doctor:
     """
-    return Doctor.query.get_or_404(doctor_id)
+    doctor = Doctor.query.get(doctor_id)
+    if doctor:
+        return doctor
+    raise RuntimeError(f"Doctor with id: {doctor_id} was not found")
 
 
 def update(doctor_id, doctor):
@@ -36,16 +39,18 @@ def update(doctor_id, doctor):
     :param doctor_id:
     :param doctor:
     """
-    doctor_for_edit = Doctor.query.get_or_404(doctor_id)
+    doctor_for_edit = Doctor.query.get(doctor_id)
+    if doctor_for_edit:
+        doctor_for_edit.full_name = doctor.full_name
+        doctor_for_edit.seniority = doctor.seniority
+        doctor_for_edit.specialty = doctor.specialty
+        doctor_for_edit.phone_number = doctor.phone_number
+        doctor_for_edit.email = doctor.email
 
-    doctor_for_edit.full_name = doctor.full_name
-    doctor_for_edit.seniority = doctor.seniority
-    doctor_for_edit.specialty = doctor.specialty
-    doctor_for_edit.phone_number = doctor.phone_number
-    doctor_for_edit.email = doctor.email
-
-    db.session.add(doctor_for_edit)
-    db.session.commit()
+        db.session.add(doctor_for_edit)
+        db.session.commit()
+    else:
+        raise RuntimeError(f"Doctor with id: {doctor_id} was not found")
 
 
 def delete(doctor_id):
@@ -53,6 +58,9 @@ def delete(doctor_id):
     Function for deleting doctor from DB
     :param doctor_id:
     """
-    doctor = Doctor.query.get_or_404(doctor_id)
-    db.session.delete(doctor)
-    db.session.commit()
+    doctor = Doctor.query.get(doctor_id)
+    if doctor:
+        db.session.delete(doctor)
+        db.session.commit()
+    else:
+        raise RuntimeError(f"Doctor with id: {doctor_id} was not found")
